requirements.txt
Django==1.8.2
Markdown==2.6.2
Pillow==2.9.0
PyYAML==3.11
arrow==0.6.0
awscli==1.9.15
awsebcli==3.6.2
backports.ssl-match-hostname==3.4.0.2
blessed==1.9.5
boto==2.38.0
botocore==1.3.15
cement==2.4.0
colorama==0.3.3
django-crispy-forms==1.4.0
django-flatpages-x==0.1.2
django-ipware==1.1.1
django-longerusernameandemail==0.5.7
django-markitup==2.2.2
django-password-reset==0.8.1
django-require==1.0.9
django-require-s3==1.0.0
django-storages==1.1.8
djangorestframework==3.2.4
djrill==2.0
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
jmespath==0.9.0
mandrill==1.0.57
pathspec==0.3.3
psycopg2==2.6.1
pyasn1==0.1.9
python-dateutil==2.4.2
pytz==2015.7
pyzillow==0.5.5
requests==2.6.2
rsa==3.2.3
six==1.10.0
stripe==1.23.0
texttable==0.8.4
wcwidth==0.1.5
websocket-client==0.34.0
wsgiref==0.1.2
